requirements.txt
boto
requests
python-simple-hipchat